Add data search custom hook
